+++ pkg/controllers/dynakube/activegate/statefulset/probe/probe.go
@@ -1,7 +1,7 @@
 // Copyright Dynatrace LLC
 // SPDX-License-Identifier: Apache-2.0
 
-package probes
+package probe
 
 import (
 	agconsts "github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/activegate/consts"
@@ -9,7 +9,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
-func BuildReadinessProbe() *corev1.Probe {
+func Readiness() *corev1.Probe {
 	return &corev1.Probe{
 		ProbeHandler: corev1.ProbeHandler{
 			HTTPGet: &corev1.HTTPGetAction{
@@ -26,7 +26,7 @@ func BuildReadinessProbe() *corev1.Probe {
 	}
 }
 
-func BuildLivenessProbe() *corev1.Probe {
+func Liveness() *corev1.Probe {
 	return &corev1.Probe{
 		ProbeHandler: corev1.ProbeHandler{
 			HTTPGet: &corev1.HTTPGetAction{
